# openapi.yaml
openapi: 3.1.0
info:
  title: UML Diagram Studio API
  description: |-
    REST API for the DiaUML Studio platform: authentication (HS512 JWT),
    user registration and account management.

    **Authentication**

    1. Call `POST /auth/login` with valid credentials to obtain a JWT.
    2. Click **Authorize** and paste the token to call protected endpoints.

    **Response envelope**

    All responses use `ApiResponse`: `{ "code": int, "message": string, "result": T }`.
    On errors `result` is absent (`@JsonInclude(NON_NULL)`). Error codes are defined
    in the `ErrorCode` enum (e.g. `1006` = invalid credentials, `9999` = unexpected error).

    **Error handling**

    All errors are handled centrally by `GlobalExceptionHandler` (@ControllerAdvice):
    - `AppException` → HTTP status from `ErrorCode.statusCode`
    - `MethodArgumentNotValidException` → 400
    - `AccessDeniedException` → 403
    - Uncaught `Exception` → 500 (code 9999)
  contact:
    name: SU26 UML Project Team
    email: uml-admin@educare.com
  license:
    name: Proprietary
  version: 0.0.1-SNAPSHOT
servers:
- url: http://localhost:8088/api/uml
  description: Local development server
security:
- bearerAuth: []
tags:
- name: Users
  description: User registration and account management.
- name: Authentication
  description: "Login, token introspection, logout and account status checks."
paths:
  /users/register:
    post:
      tags:
      - Users
      summary: Register a new user account
      description: "Creates a new user after validating field constraints, username/email\
        \ uniqueness, and the existence of the assigned role."
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterRequest"
            example:
              username: johndoe
              password: Passw0rd
              fullName: John Doe
              email: john.doe@gmail.com
              phone: "0901234567"
              roleId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        required: true
      responses:
        "200":
          description: User registered successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                code: 200
                message: Đăng ký thành công
                result:
                  userID: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  username: johndoe
                  fullName: John Doe
                  email: john.doe@gmail.com
                  phone: "0901234567"
                  status: ACTIVE
                  role:
                    roleID: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                    roleName: USER
                    description: Standard application user
        "400":
          description: "Validation failure or bad request — `MethodArgumentNotValidException`\
            \ or `AppException` with a 4xx `ErrorCode` (e.g. `USER_EXISTED` 1002,\
            \ `INVALID_PASSWORD` 1005)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: "Not authenticated — missing, expired, or invalid JWT. Produced\
            \ by `JwtAuthenticationEntryPoint` (code 401) or `AppException(UNAUTHENTICATED)`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "403":
          description: Authenticated but not authorised — `AccessDeniedException`
            or `AppException(USER_INACTIVE)` (code 403).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "500":
          description: Unexpected server error — unhandled `Exception` fallback (code
            9999).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
      security: []
  /auth/logout:
    post:
      tags:
      - Authentication
      summary: Log out (invalidate a JWT)
      description: Blacklists the supplied token until its natural expiry so it can
        no longer be used.
      operationId: logout
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
            example:
              token: eyJhbGciOiJIUzUxMiJ9...
        required: true
      responses:
        "200":
          description: Token invalidated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                code: 0
        "400":
          description: "Validation failure or bad request — `MethodArgumentNotValidException`\
            \ or `AppException` with a 4xx `ErrorCode` (e.g. `USER_EXISTED` 1002,\
            \ `INVALID_PASSWORD` 1005)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: "Not authenticated — missing, expired, or invalid JWT. Produced\
            \ by `JwtAuthenticationEntryPoint` (code 401) or `AppException(UNAUTHENTICATED)`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "403":
          description: Authenticated but not authorised — `AccessDeniedException`
            or `AppException(USER_INACTIVE)` (code 403).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "500":
          description: Unexpected server error — unhandled `Exception` fallback (code
            9999).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
      security: []
  /auth/login:
    post:
      tags:
      - Authentication
      summary: Authenticate and obtain a JWT
      description: Validates username/password and returns a signed HS512 JWT (valid
        for 1 hour).
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              username: johndoe
              password: Passw0rd
        required: true
      responses:
        "200":
          description: Authentication succeeded; JWT returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                code: 200
                result:
                  token: eyJhbGciOiJIUzUxMiJ9...
                  authenticated: true
        "400":
          description: "Validation failure or bad request — `MethodArgumentNotValidException`\
            \ or `AppException` with a 4xx `ErrorCode` (e.g. `USER_EXISTED` 1002,\
            \ `INVALID_PASSWORD` 1005)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: "Not authenticated — missing, expired, or invalid JWT. Produced\
            \ by `JwtAuthenticationEntryPoint` (code 401) or `AppException(UNAUTHENTICATED)`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "403":
          description: Authenticated but not authorised — `AccessDeniedException`
            or `AppException(USER_INACTIVE)` (code 403).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "500":
          description: Unexpected server error — unhandled `Exception` fallback (code
            9999).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
      security: []
  /auth/introspect:
    post:
      tags:
      - Authentication
      summary: Introspect (validate) a JWT
      description: "Returns whether the supplied token is structurally valid, correctly\
        \ signed, unexpired and not blacklisted."
      operationId: introspect
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntrospectRequest"
            example:
              token: eyJhbGciOiJIUzUxMiJ9...
        required: true
      responses:
        "200":
          description: "Introspection result (valid: true|false)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                code: 200
                result:
                  valid: true
        "400":
          description: "Validation failure or bad request — `MethodArgumentNotValidException`\
            \ or `AppException` with a 4xx `ErrorCode` (e.g. `USER_EXISTED` 1002,\
            \ `INVALID_PASSWORD` 1005)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: "Not authenticated — missing, expired, or invalid JWT. Produced\
            \ by `JwtAuthenticationEntryPoint` (code 401) or `AppException(UNAUTHENTICATED)`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "403":
          description: Authenticated but not authorised — `AccessDeniedException`
            or `AppException(USER_INACTIVE)` (code 403).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "500":
          description: Unexpected server error — unhandled `Exception` fallback (code
            9999).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
      security: []
  /auth/account-status:
    get:
      tags:
      - Authentication
      summary: Check whether an account is locked
      description: "Looks up a user by username or email and reports whether the account\
        \ is in the LOCKED state. Always returns 200 — unknown identifiers are reported\
        \ as locked: false to avoid user enumeration."
      operationId: accountStatus
      parameters:
      - name: identifier
        in: query
        description: Username or email address to look up.
        required: true
        schema:
          type: string
        example: johndoe
      responses:
        "200":
          description: Lock status returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                code: 0
                result:
                  locked: false
        "400":
          description: "Validation failure or bad request — `MethodArgumentNotValidException`\
            \ or `AppException` with a 4xx `ErrorCode` (e.g. `USER_EXISTED` 1002,\
            \ `INVALID_PASSWORD` 1005)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: "Not authenticated — missing, expired, or invalid JWT. Produced\
            \ by `JwtAuthenticationEntryPoint` (code 401) or `AppException(UNAUTHENTICATED)`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "403":
          description: Authenticated but not authorised — `AccessDeniedException`
            or `AppException(USER_INACTIVE)` (code 403).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "500":
          description: Unexpected server error — unhandled `Exception` fallback (code
            9999).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
      security: []
components:
  schemas:
    UserRegisterRequest:
      type: object
      description: Payload for registering a new user account.
      properties:
        username:
          type: string
          description: "Unique username, 3-100 characters."
          example: johndoe
          maxLength: 100
          minLength: 3
        password:
          type: string
          description: "Min 6 chars, must contain at least one letter and one digit."
          example: Passw0rd
          maxLength: 2147483647
          minLength: 6
          pattern: "^(?=.*[A-Za-z])(?=.*\\d)[A-Za-z\\d@$!%*#?&]{6,}$"
        fullName:
          type: string
          description: "Display name, up to 255 characters."
          example: John Doe
          maxLength: 255
          minLength: 0
        email:
          type: string
          description: "Unique, valid email address."
          example: john.doe@gmail.com
          minLength: 1
          pattern: "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,6}$"
        phone:
          type: string
          description: "Phone number, 10-11 digits."
          example: "0901234567"
          pattern: "^[0-9]{10,11}$"
        roleId:
          type: string
          format: uuid
          description: UUID of the role to assign to the user.
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
      required:
      - email
      - fullName
      - password
      - roleId
      - username
    ApiResponse:
      type: object
      description: Standard response envelope wrapping every endpoint result.
      properties:
        code:
          type: integer
          format: int32
          description: Application status code (200 = success; see ErrorCode enum
            for errors).
          example: 200
        message:
          type: string
          description: Human-readable message.
          example: Success
        result:
          description: Payload of the response; null when there is no body.
    LogoutRequest:
      type: object
      description: Wraps the JWT to invalidate.
      properties:
        token:
          type: string
          description: The JWT to invalidate (blacklist until expiry).
          example: eyJhbGciOiJIUzUxMiJ9...
      required:
      - token
    LoginRequest:
      type: object
      description: Credentials used to obtain a JWT.
      properties:
        username:
          type: string
          description: Account username.
          example: johndoe
          minLength: 1
        password:
          type: string
          description: Account password.
          example: Passw0rd
          minLength: 1
      required:
      - password
      - username
    IntrospectRequest:
      type: object
      description: Wraps a JWT to be validated.
      properties:
        token:
          type: string
          description: The JWT to introspect.
          example: eyJhbGciOiJIUzUxMiJ9...
      required:
      - token
  securitySchemes:
    bearerAuth:
      type: http
      description: |-
        Paste the JWT returned by `POST /auth/login` (the `result.token` field).
        Swagger sends it as the `Authorization: Bearer <token>` header on protected endpoints.
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT